♻️ Refactor
- type 위치 변경
- button type 정의
등으로 인한 리팩토링
