+++ src/Page/ResultPage/index.tsx
@@ -1,7 +1,8 @@
 import { useLocation } from 'react-router-dom';
+import ProblemTitle from '../../Component/Utils/ProblemTitle';
 import { listData } from '../../data';
 import Header from '../../Template/Header';
-import { IProblemIdLinkState } from '../../types';
+import { IProblemIdLinkState } from '../../types/problem';
 
 function ResultPage() {
   const state = useLocation().state as IProblemIdLinkState;
@@ -10,7 +11,18 @@ function ResultPage() {
   return (
     <>
       <Header />
-      <div></div>
+      <div>
+        <ProblemTitle
+          id={problemData.id}
+          title={problemData.title}
+          numberSolved={0}
+          averageScore={0}
+          highestScore={0}
+          lowestScore={0}
+          tagList={[]}
+          desc={''}
+        />
+      </div>
     </>
   );
 }
