+++ src/types/util.ts
@@ -1,0 +1,6 @@
+interface IDropdownElement {
+  id: string;
+  name: string;
+}
+
+export type { IDropdownElement };
